chore: rename utils to context_auth
