+++ src/servers/src/grpc.rs
@@ -14,13 +14,13 @@
 
 pub mod builder;
 mod cancellation;
+pub mod context_auth;
 mod database;
 pub mod flight;
 pub mod frontend_grpc_handler;
 pub mod greptime_handler;
 pub mod prom_query_gateway;
 pub mod region_server;
-pub mod utils;
 
 use std::net::SocketAddr;
